refactor: reconfigure design
