+++ server/model/player.py
@@ -1,15 +1,8 @@
 from dataclasses import dataclass, field
 from typing import List
 
+from model.call import Call
 from model.tile import Tile
-
-
-@dataclass
-class Call:
-    type: str
-    tiles: List[Tile]
-    target_player_id: int
-    target_tile_id: int
 
 
 @dataclass
