+++ server/presenter/controller/room.py
@@ -1,7 +1,7 @@
 from socketio import Server
 
-import usecase.room
-import usecase.player
+import application.usecase.room as room_usecase
+import application.usecase.player as player_usecase
 import presenter.response.room as emit
 
 
@@ -16,18 +16,18 @@ def set(socket_io: Server):
         #     message: str = "名前を入力してください"
         #     emit.notify_error(socket_io, [socket_id], message)
         #     return
-        player = usecase.player.register_player(player_name, socket_id)
-        room = usecase.room.create_room(player.id)
-        usecase.room.enter_room(room.number, player.id)
+        player = player_usecase.register_player(player_name, socket_id)
+        room = room_usecase.create_room(player.id)
+        room_usecase.enter_room(room.number, player.id)
 
         emit.entered_room(socket_io, [socket_id], room.number)
 
     @socket_io.on("enter_room")
     def on_enter_room(socket_id: str, player_name: str, room_number_str: str):
         room_number = int(room_number_str)
-        player = usecase.player.register_player(player_name, socket_id)
-        usecase.room.enter_room(room_number, player.id)
-        players = usecase.room.get_players_in_room(room_number)
+        player = player_usecase.register_player(player_name, socket_id)
+        room_usecase.enter_room(room_number, player.id)
+        players = room_usecase.get_players_in_room(room_number)
 
         emit.entered_room(socket_io, [socket_id], room_number)
         emit.joined_room(socket_io,
@@ -36,10 +36,10 @@ def set(socket_io: Server):
 
     @socket_io.on("leave_room")
     def on_leave_room(socket_id: str):
-        player = usecase.room.unready(socket_id)
-        room = usecase.room.get_room_by_player_id(player.id)
-        usecase.room.leave_room(room.id, player.id)
-        players = usecase.room.get_players_in_room(room.number)
+        player = room_usecase.unready(socket_id)
+        room = room_usecase.get_room_by_player_id(player.id)
+        room_usecase.leave_room(room.id, player.id)
+        players = room_usecase.get_players_in_room(room.number)
 
         emit.left_room(socket_io,
                        [p.socket_id for p in players],
@@ -47,17 +47,17 @@ def set(socket_io: Server):
 
     @socket_io.on("connect_waiting_room")
     def on_connect_waiting_room(new_socket_id: str, old_socket_id: str):
-        usecase.room.reconnect(new_socket_id, old_socket_id)
-        player_names = usecase.room.get_players(new_socket_id)
+        room_usecase.reconnect(new_socket_id, old_socket_id)
+        player_names = room_usecase.get_players(new_socket_id)
 
         emit.reconnected(socket_io, [new_socket_id], new_socket_id)
         emit.players_info(socket_io, [new_socket_id], player_names)
 
     @socket_io.on("ready_game")
     def on_ready_game(socket_id: str):
-        player = usecase.room.ready(socket_id)
-        room = usecase.room.get_room_by_player_id(player.id)
-        players = usecase.room.get_players_in_room(room.number)
+        player = room_usecase.ready(socket_id)
+        room = room_usecase.get_room_by_player_id(player.id)
+        players = room_usecase.get_players_in_room(room.number)
 
         emit.readied_game(socket_io,
                           [p.socket_id for p in players],
@@ -65,9 +65,9 @@ def set(socket_io: Server):
 
     @socket_io.on("unready_game")
     def on_unready_game(socket_id: str):
-        player = usecase.room.unready(socket_id)
-        room = usecase.room.get_room_by_player_id(player.id)
-        players = usecase.room.get_players_in_room(room.number)
+        player = room_usecase.unready(socket_id)
+        room = room_usecase.get_room_by_player_id(player.id)
+        players = room_usecase.get_players_in_room(room.number)
 
         emit.unreadied_game(socket_io,
                             [p.socket_id for p in players],
@@ -75,8 +75,8 @@ def set(socket_io: Server):
 
     @socket_io.on("start_game")
     def on_start_game(socket_id: str):
-        player = usecase.room.unready(socket_id)
-        room = usecase.room.get_room_by_player_id(player.id)
-        players = usecase.room.get_players_in_room(room.number)
+        player = room_usecase.unready(socket_id)
+        room = room_usecase.get_room_by_player_id(player.id)
+        players = room_usecase.get_players_in_room(room.number)
 
         emit.started_game(socket_io, [p.socket_id for p in players])
