--- server/repository/score.py
@@ -1,3 +1,0 @@
-from typing import Tuple
-
-from db.database import execute_query, fetch_data
